Crate ril

The Rust Imaging Library. A high-level image processing crate for Rust.

Modules

draw

encode

encodings

error

image

pixel

prelude

The crate prelude exports. Importing this with a wildcard will import most items from RIL that can be useful for image processing, along with bringing crucial traits into scope.

sequence

Structs

BitPixel

Represents a single-bit pixel that represents either a pixel that is on or off.

Border

Represents a shape border.

Ellipse

An ellipse, which could be a circle.

Frame

Represents a frame in an image sequence. It encloses an Image and extra metadata about the frame.

Image

A high-level image representation.

ImageSequence

Represents a sequence of image frames such as an animated image.

L

Represents an L, or luminance pixel that is stored as only one single number representing how bright, or intense, the pixel is.

Paste

Pastes or overlays an image on top of another image.

Rectangle

A rectangle.

Rgb

Represents an RGB pixel.

Rgba

Represents an RGBA pixel.

Enums

BorderPosition

Represents whether a border is inset, outset, or if it lays in the center.

DisposalMethod

The method used to dispose a frame before transitioning to the next frame in an image sequence.

Dynamic

Represents a pixel type that is dynamically resolved.

DynamicFrameIterator

Represents any one of the different types of frame iterators, compacted into one common enum with common methods.

Error

Represents an error that occurs within the crate.

ImageFormat

Represents the underlying encoding format of an image.

LoopCount

OverlayMode

The behavior to use when overlaying images on top of each other.

ResizeAlgorithm

A filtering algorithm that is used to resize an image.

Traits

Alpha

Represents a pixel that supports alpha, or transparency values.

Banded

Represents an image with multiple channels, called bands.

Decoder

Low-level decoder interface around an image format.

Draw

Encoder

Low-level encoder interface around an image format.

FrameIterator

Represents the lazy decoding of frames from an encoded image sequence, such as an animated image.

Pixel

Represents any type of pixel in an image.

Type Definitions

Result