ril

Module draw

Encloses most drawing implementations and drawable objects.

Structs

Border

Represents a shape border.

Ellipse

An ellipse, which could be a circle.

Line

A line.

Paste

Pastes or overlays an image on top of another image.

Polygon

A polygon.

Rectangle

A rectangle.

Enums

BorderPosition

Represents whether a border is inset, outset, or if it lays in the center.

Traits

Draw

A common trait for all objects able to be drawn on an image.