Struct ril::draw::Line

pub struct Line<F: IntoFill> {
    pub color: F::Fill,
    pub mode: Option<OverlayMode>,
    pub thickness: u32,
    pub start: (u32, u32),
    pub end: (u32, u32),
    pub antialiased: bool,
    pub rounded: bool,
    pub position: BorderPosition,
}

A line.

At its core, this method utilizes Bresenhamโs line algorithm, and Xiaolin Wuโs line algorithm for antialiased lines. Thicker lines are drawn as polygons.

Fields
color: F::Fill

The color of the line.

mode: Option<OverlayMode>

The overlay mode of the line, or None to inherit from the overlay mode of the image.

thickness: u32

The thickness of the line, in pixels. Defaults to 1.

start: (u32, u32)

The start point of the line.

end: (u32, u32)

The end point of the line.

antialiased: bool

Whether the line should be antialiased. Note that drawing antialiased lines is slower than drawing non-antialiased lines. Defaults to false.

rounded: bool

Whether the endpoints of the line should be โrounded offโ with circles. Defaults to false. Currently, endpoints are not antialiased, regardless of the value of antialiased.

Note that for even-numbered thicknesses, the endpoints will not be perfectly aligned with the line. For optimal results, use odd-numbered thicknesses when using enabling this field.

position: BorderPosition

The position of the line relative to the start and end points. Defaults to Center (which is different from the default of Border).

Implementations

impl<F: IntoFill> Line<F>

pub fn new(start: (u32, u32), end: (u32, u32), color: F) -> Self

Creates a new line.

pub fn with_color(self, color: F) -> Self

Sets the color of the line.

pub const fn with_mode(self, mode: OverlayMode) -> Self

Sets the overlay mode of the line.

pub const fn with_thickness(self, thickness: u32) -> Self

Sets the thickness of the line.

pub fn with_start(self, x: u32, y: u32) -> Self

Sets the start coordinates of the line.

pub fn with_end(self, x: u32, y: u32) -> Self

Sets the end coordinates of the line.

pub const fn with_antialiased(self, antialiased: bool) -> Self

Sets whether the line should be antialiased. If this is set to true, the overlay mode of this line will also be set to OverlayMode::Merge.

pub const fn with_rounded(self, rounded: bool) -> Self

Sets whether the line should be rounded.

pub const fn with_position(self, position: BorderPosition) -> Self

Sets the position of the line relative to the start and end points.