ril::draw

Trait Draw

pub trait Draw<P: Pixel> {
    // Required method
    fn draw<I: DerefMut<Target = Image<P>>>(&self, image: I);
}

A common trait for all objects able to be drawn on an image.

Whether or not to implement this trait is more or less a matter of semantics.

Required Methods

fn draw<I: DerefMut<Target = Image<P>>>(&self, image: I)

Draws the object to the given image.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.