ril::encode

Trait Encoder

pub trait Encoder<P: Pixel, W: Write>: Sized {
    type Config: Default;

    // Required methods
    fn new(
        dest: W,
        metadata: impl HasEncoderMetadata<Self::Config, P>,
    ) -> Result<Self>;
    fn add_frame(&mut self, frame: &impl FrameLike<P>) -> Result<()>;
    fn finish(self) -> Result<()>;

    // Provided methods
    fn encode_static(image: &Image<P>, dest: W) -> Result<()> { ... }
    fn encode_sequence(sequence: &ImageSequence<P>, dest: W) -> Result<()> { ... }
}

Low-level encoder interface around an image format. Typically only accessed for lazy encoding.

Encoder metadata

When preparing an encoder, you must pass in some metadata to the encoder. This metadata is used to configure the encoder for encoding the image. The metadata can be:

A static Image, in which case the encoder will be prepared for encoding a static image.

A Frame, in which case the encoder will be prepared for encoding an image sequence.

An ImageSequence, in which case the encoder will be prepared for encoding an image sequence.

An EncoderMetadata object for more fine-grained control over the encoder.

Notes

The specific metadata that is actually used and/or considered depends on the encoder implementation.

If you provide an image, frame, or sequence as metadata, you are only registering metadata, the image or frame itself will not be encoded yet. You would have to make a call to [Self::encode_frame] to encode the image or frame.

Use EncoderMetadata::from to create encoder metadata from an image, frame, or sequence.

Examples

use ril::encodings::png::{PngEncoder, PngEncoderOptions};

// Create a new encoder with default options given the width, height, and pixel type:
let encoder = PngEncoder::new(writer, EncoderMetadata::<Rgba>::new(256, 256));

// Create a new encoder will encode an animated image with 2 frames, looping infinitely:
let metadata = EncoderMetadata::<Rgba>::new(256, 256).with_sequence(2, LoopCount::Infinite);
let encoder = PngEncoder::new(writer, metadata);

// Using an image/frame/sequence as metadata:
let image = Image::<Rgb>::open("sample.png")?; // can be any Image, Frame, or ImageSequence
let encoder = PngEncoder::new(writer, &image);

// Using a frame as baseline metadata but explicitly anticipating frame count
// NOTE: See "Anticipating the frame and loop counts" below for more information
let frame = Frame::from_image(image);
let metadata = EncoderMetadata::from(&frame).with_sequence(2, LoopCount::Infinite);
let encoder = PngEncoder::new(writer, metadata);

Anticipating frame and loop counts

All frame and loops counts for the following encoders must be anticipated prior to encoding:

[PngEncoder]

This is a limitation caused by the way these encoders/codecs are implemented. For example, in the PNG format, the frame and loop counts are stored in the header of the file, which is written at the very beginning of encoding.

You can anticipate the frame and loop counts by using EncoderMetadata::with_sequence. If you have a reference to an image, frame, or sequence, you can use EncoderMetadata::from to create encoder metadata from it:

use ril::encodings::png::{PngEncoder, PngEncoderOptions};

// Grab our image, frame, or image sequence:
let frame = Frame::from_image(Image::<Rgb>::open("sample.png")?);
// Convert it into encoder metadata with `EncoderMetadata::from`,
// then anticipate the frame and loop counts with `EncoderMetadata::with_sequence`:
let metadata = EncoderMetadata::from(&frame).with_sequence(2, LoopCount::Infinite);
// Finally, create the encoder with our modified metadata:
let encoder = PngEncoder::new(writer, metadata);

If you are not certain how many frames or loops your image will have, you can use an upper bound for the frame count or loop count.

Example

Lazily encode images into an animated PNG (APNG):

use std::fs::File;
use std::time::Duration;
use ril::colors::{RED, BLUE};
use ril::encodings::png::PngEncoder;
use ril::prelude::*;

const HALF_SECOND: Duration = Duration::from_millis(500);

fn main() -> ril::Result<()> {
    let mut writer = File::create("output.png")?;
    // Prepare metadata. The PNG encoder requires we anticipate the frame count prior to encoding:
    let metadata = EncoderMetadata::<Rgb>::new(100, 100).with_frame_count(2);
    // Create the encoder
    let mut encoder = PngEncoder::new(&mut writer, metadata)?;

    // Frame 1: 100x100 red square for 0.5s
    let image = Image::new(100, 100, RED);
    encoder.add_frame(&Frame::from_image(image).with_delay(HALF_SECOND))?;

    // Frame 2: 100x100 blue square for 0.5s
    let image = Image::new(100, 100, BLUE);
    encoder.add_frame(&Frame::from_image(image).with_delay(HALF_SECOND))?;

    // Finish encoding
    encoder.finish()
}

Required Associated Types

type Config: Default

The type of the configuration object used to configure options for the encoder.

Required Methods

fn new( dest: W, metadata: impl HasEncoderMetadata<Self::Config, P>, ) -> Result<Self>

Creates a new encoder and prepares it for encoding an image.

This method is only used for lazy encoding of image sequences.

If you are encoding a static image, use Self::encode_static instead. There is no need to call any intermediate method for encoding static images since they inherently cannot be lazily encoded after all.

If you are encoding an ImageSequence directly, use Self::encode_sequence instead. Since ImageSequence already contains all the frames, there is no need to lazily encode them.

Metadata

The metadata parameter is used to provide initial metadata to the encoder. See the trait-level documentation for more information.

Notes

The specific metadata that is actually used and/or considered depends on the encoder implementation.

If you provide an image or frame as metadata, you are only registering metadata, the image or frame itself will not be encoded yet. You would have to make a call to [Self::encode_frame] to encode the image or frame.

Errors

An error occured during encoding.

The image is invalid for encoding.

fn add_frame(&mut self, frame: &impl FrameLike<P>) -> Result<()>

Adds a frame to the encoding sequence and encodes it into the given writer.

This method is only used for lazy encoding of image sequences.

If you are encoding a static image, use Self::encode_static instead. There is no need to call any intermediate method for encoding static images since they inherently cannot be lazily encoded after all.

If you are encoding an ImageSequence directly, use Self::encode_sequence instead. Since ImageSequence already contains all the frames, there is no need to lazily encode them.

Errors

An error occured during encoding.

fn finish(self) -> Result<()>

Finishes encoding the image. This must be called once after all frames are encoded.

Errors

An error occured during encoding or writing.

Provided Methods

fn encode_static(image: &Image<P>, dest: W) -> Result<()>

Encodes a static image into the given writer.

Errors

An error occured during encoding.

fn encode_sequence(sequence: &ImageSequence<P>, dest: W) -> Result<()>

Encodes the given image sequence into the given writer.

Errors

An error occured during encoding.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.