Module ril::encodings

Contains encoder and decoder implementations for various image formats.

Modules

gif

jpeg

png

webp

Enums

ColorType

Represents an arbitrary color type. Note that this does not store the bit-depth or the type used to store the value of each channel, although it can specify the number of channels.