Struct ril::encodings::gif::GifEncoder

pub struct GifEncoder { /* private fields */ }

A GIF encoder interface over [gif::Encoder].

Implementations

impl GifEncoder

pub const fn new() -> Self

Creates a new encoder with default settings.

pub fn with_speed(self, speed: u8) -> Self

Sets the speed of the encoder. Higher speeds come at the cost of lower image quality. Must be between 1 and 30.

Panics

Speed is not between 1 and 30.

pub const unsafe fn with_speed_unchecked(self, speed: u8) -> Self

Sets the speed of the encoder. Higher speeds come at the cost of lower image quality. Should be between 1 and 30, but this doesn’t check for that.

Safety

Make sure speed is between 1 and 30.

See Also

[with_speed] for the safe, checked version of this method.