Module ril::encodings::png

Structs

ApngFrameIterator

PngDecoder

A PNG decoder interface around [png::Decoder].

PngEncoder

A PNG encoder interface around [png::Encoder].

PngEncoderOptions

PNG configuration options for PngEncoder.

Enums

AdaptiveFilterType

The filtering method for preprocessing scanline data before compression.

Compression

The type and strength of applied compression.

FilterType

The byte level filter applied to scanlines to prepare them for compression.