Enum ril::ImageFormat

pub enum ImageFormat {
    Unknown,
    Png,
    Jpeg,
    Gif,
    Bmp,
    Tiff,
    WebP,
}

Represents the underlying encoding format of an image.

Variants

Unknown

No known encoding is known for the image.

This is usually because the image was created manually. See Image::set_format to manually set the encoding format.

Png

The image is encoded in the PNG format.

Jpeg

The image is encoded in the JPEG format.

Gif

The image is encoded in the GIF format.

Bmp

The image is encoded in the BMP format.

Tiff

The image is encoded in the TIFF format.

WebP

The image is encoded in the WebP format.

Implementations

impl ImageFormat

pub fn is_unknown(&self) -> bool

Returns whether the extension is unknown.

pub fn from_extension(ext: impl AsRef<OsStr>) -> Result<Self>

Parses the given extension and returns the corresponding image format.

If the extension is an unknown extension, Ok(ImageFormat::Unknown) is returned.

If the extension is completely invalid and fails to be converted into a &str, the InvalidExtension error is returned.

Errors

The extension is completely invalid and failed to be converted into a &str.

pub fn from_path(path: impl AsRef<Path>) -> Result<Self>

Returns the format specified by the given path.

This uses ImageFormat::from_extension to parse the extension.

This resolves via the extension of the path. See ImageFormat::infer_encoding for an implementation that can resolve the format from the data.

Errors

No extension can be resolved from the path.

pub fn from_mime_type(mime: impl AsRef<str>) -> Self

Returns the format specified by the given MIME type.

pub fn infer_encoding(sample: &[u8]) -> Self

Infers the encoding format from the given data via a byte stream.

pub fn run_encoder<P: Pixel>( &self, image: &Image<P>, dest: impl Write, ) -> Result<()>

Encodes the Image into raw bytes.

Errors

An error occured while encoding.

Panics

No encoder implementation is found for this image encoding.

pub fn run_sequence_encoder<P: Pixel>( &self, seq: &ImageSequence<P>, dest: impl Write, ) -> Result<()>

Encodes the ImageSequence into raw bytes. If the encoding does not supported image sequences (or multi-frame images), it will only encode the first frame.

Errors

An error occured while encoding.

Panics

No encoder implementation is found for this image encoding.

pub fn run_decoder<P: Pixel>(&self, stream: impl Read) -> Result<Image<P>>

Decodes the image data from into an image.

Errors

An error occured while decoding.

Panics

No decoder implementation is found for this image encoding.

pub fn run_sequence_decoder<'a, P: Pixel + 'a, R: Read + 'a>( &self, stream: R, ) -> Result<Box<dyn FrameIterator<P> + 'a>>

Decodes the image sequence data into an image sequence.

Errors

An error occured while decoding.

Panics

No decoder implementation is found for this image encoding.