Struct ril::pixel::NoOp

pub struct NoOp;

A pixel type that does and stores nothing. This pixel type is useless and will behave weirdly with your code. This is usually only used for internal or polyfill purposes.