Trait ril::pixel::Alpha

pub trait Alpha: Pixel {
    // Required methods
    fn alpha(&self) -> u8;
    fn with_alpha(self, alpha: u8) -> Self;
}

Represents a pixel that supports alpha, or transparency values.

Required Methods

fn alpha(&self) -> u8

Returns the alpha, or opacity level of the pixel.

This is a value between 0 and 255. 0 is completely transparent, and 255 is completely opaque.

fn with_alpha(self, alpha: u8) -> Self

Clones this pixel with the given alpha value.

Object Safety

This trait is not object safe.