Trait ril::pixel::Paletted

pub trait Paletted<'p>: Pixelwhere
    Self: 'p,{
    fn from_palette(palette: &'p [Self::Color], index: Self::Subpixel) -> Self;
    fn palette(&self) -> &'p [Self::Color];
    fn palette_index(&self) -> Self::Subpixel;

    fn color(&self) -> Self::Color { ... }
    unsafe fn color_unchecked(&self) -> Self::Color { ... }
}

A trait representing a paletted pixel. Pixel::Subpixel is the type of the palette index.

The generic lifetime parameter 'p represents the lifetime of a palette the type will hold a reference to.

Required Methods

fn from_palette(palette: &'p [Self::Color], index: Self::Subpixel) -> Self

Creates this pixel from the given palette and index. For unpaletted pixels, use Pixel::from_arbitrary_palette instead.

Errors

The pixel index is invalid/out of bounds.

If the color type is not supported by the pixel type.

An error occurs when trying to convert the data to the pixel type.

fn palette(&self) -> &'p [Self::Color]

Returns the palette lookup as a slice.

fn palette_index(&self) -> Self::Subpixel

Returns the index in the palette this pixel is of.

Provided Methods

fn color(&self) -> Self::Color

Resolves the color of the pixel. Because invalid palette values are supposed to be propagated prior to calling this, this will panic

unsafe fn color_unchecked(&self) -> Self::Color

Resolves the color of the pixel. Invalid palette values should be propagated prior to to calling this, but it isn’t guaranteed, for example if the palette pixel was manually initialized.

This results in undefined behavior if the palette index is invalid.

Safety

The palette index must be valid.