Struct ril::sequence::ImageSequence

pub struct ImageSequence<P: Pixel> { /* private fields */ }

Represents a sequence of image frames such as an animated image.

See Image for the static image counterpart, and see Frame to see how each frame is represented in an image sequence.

Implementations

impl<P: Pixel> ImageSequence<P>

pub fn new() -> Self

Creates a new image sequence with no frames.

Note

A frameless image sequence is forbidden to be encoded and you will receive a panic.

pub fn from_reader<R: Read>(
format: ImageFormat,
bytes: R
) -> Result<DynamicFrameIterator<P, R>>

Decodes the image sequence with the explicitly given image encoding from the raw byte reader.

This decodes frames lazily as an iterator. Call [DynamicFrameIterator::into_sequence] to collect all frames greedily into an ImageSequence.

If the image sequence is a single-frame static image or if the encoding format does not support animated images, this will just return an image sequence containing one frame.

Errors

DecodingError: The image could not be decoded, maybe it is corrupt.

pub fn from_reader_inferred<R: Read + Write>(
bytes: R
) -> Result<DynamicFrameIterator<P, R>>

Decodes an image sequence from the given read stream of bytes, inferring its encoding.

This decodes frames lazily as an iterator. Call [DynamicFrameIterator::into_sequence] to collect all frames greedily into an ImageSequence.

If the image sequence is a single-frame static image or if the encoding format does not support animated images, this will just return an image sequence containing one frame.

Note

The bound on bytes includes Write due to a Rust limitation. This will be looked into in the future to not require Write.

If you are limited by this trait bound, you can either specify the image format manually using [from_reader], or you can try using ImageFormat::infer_encoding along with [from_reader] manually instead. If you are able to use [from_bytes] instead, which takes a byte slice instead of a Read stream, you can either that or [from_bytes_inferred], too, which does not require a Write bound either.

Errors

DecodingError: The image could not be decoded, maybe it is corrupt.

UnknownEncodingFormat: Could not infer the encoding from the image. Try explicitly specifying it.

Panics

No decoder implementation for the given encoding format.

pub fn from_bytes(
format: ImageFormat,
bytes: &[u8]
) -> Result<DynamicFrameIterator<P, &[u8]>>

Decodes an image sequence with the explicitly given image encoding from the byte slice. Could be useful in conjunction with the include_bytes! macro.

Currently, this is not any different than [from_reader].

This decodes frames lazily as an iterator. Call [DynamicFrameIterator::into_sequence] to collect all frames greedily into an ImageSequence.

If the image sequence is a single-frame static image or if the encoding format does not support animated images, this will just return an image sequence containing one frame.

Note

This takes different parameters than Image::from_bytes - that takes any AsRef<[u8]> while this strictly only takes byte slices (&[u8]).

Errors

DecodingError: The image could not be decoded, maybe it is corrupt.

Panics

No decoder implementation for the given encoding format.

pub fn from_bytes_inferred(
bytes: &[u8]
) -> Result<DynamicFrameIterator<P, &[u8]>>

Decodes an image sequence from the given byte slice, inferring its encoding. Could be useful in conjunction with the include_bytes! macro.

This is more efficient than [from_reader_inferred], and can act as a workaround for bypassing the Write trait bound.

This decodes frames lazily as an iterator. Call [DynamicFrameIterator::into_sequence] to collect all frames greedily into an ImageSequence.

If the image sequence is a single-frame static image or if the encoding format does not support animated images, this will just return an image sequence containing one frame.

Note

This takes different parameters than Image::from_bytes - that takes any AsRef<[u8]> while this strictly only takes byte slices (&[u8]).

Errors

DecodingError: The image could not be decoded, maybe it is corrupt.

UnknownEncodingFormat: Could not infer the encoding from the image. Try explicitly specifying it.

Panics

No decoder implementation for the given encoding format.

pub fn open(path: impl AsRef<Path>) -> Result<DynamicFrameIterator<P, File>>

Opens a file from the given path and decodes it, returning an iterator over its frames.

The encoding of the image is automatically inferred. You can explicitly pass in an encoding by using the [from_reader] method.

Note

Unlike the inference of Image::open this does not infer from raw bytes if inferring from file extension fails; instead it immediately returns the error.

Errors

todo!()

pub fn encode(&self, encoding: ImageFormat, dest: &mut impl Write) -> Result<()>

Encodes this image sequence with the given encoding and writes it to the given write buffer.

Errors

An error occured during encoding.

Panics

No encoder implementation for the given encoding format.

pub fn save(&self, encoding: ImageFormat, path: impl AsRef<Path>) -> Result<()>

Saves the image sequence with the given encoding to the given path. You can try saving to a memory buffer by using the [encode] method.

Errors

An error occured during encoding.

Panics

No encoder implementation for the given encoding format.

pub fn save_inferred(&self, path: impl AsRef<Path>) -> Result<()>

Saves the image sequence to the given path, inferring the encoding from the given path/filename extension.

This is obviously slower than [save] since this method itself uses it. You should only use this method if the filename is dynamic, or if you do not know the desired encoding before runtime.

See [save] for more information on how saving works.

Errors

Could not infer encoding format.

An error occured during encoding.

Panics

No encoder implementation for the given encoding format.

pub fn from_frames(frames: Vec<Frame<P>>) -> Self

Creates a new image sequence from the given frames

pub fn with_frame(self, frame: Frame<P>) -> Self

Adds a new frame to this image sequence and returns this sequence. Useful for method-chaining.

pub fn push_frame(&mut self, frame: Frame<P>)

Adds a new frame to this image sequence.

pub fn extend_frames<I>(&mut self, frames: I) where
I: IntoIterator<Item = Frame<P>>,

Extends frames from the given iterator.

pub const fn loop_count(&self) -> LoopCount

Returns how many times this image sequence loops for.

pub const fn with_loop_count(self, loops: LoopCount) -> Self

Takes this image and sets how many times this image sequence loops for.

pub fn set_loop_count(&mut self, loops: LoopCount)

Sets how many times this image sequence loops for in place.

pub const fn looped_exactly(self, loops: u32) -> Self

Sets the exact number of loops this image sequence loops for.

pub const fn looped_infinitely(self) -> Self

Sets the image sequence to loop infinitely.

pub fn into_frames(self) -> Vec<Frame<P>>

Consumes this image sequence and returns the frames it contains.

pub fn iter(&self) -> impl Iterator<Item = &Frame<P>>

Iterates through the frames in this image sequence by reference.

pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut Frame<P>>

Iterates through the frames in this image sequence by mutable reference.

pub fn is_empty(&self) -> bool

Returns whether there are no frames in the image sequence. If so, this will probably be invalid to encode.

pub fn len(&self) -> usize

Returns the number of frames in this image sequence.

pub fn into_first_image(self) -> Image<P>

Consumes this image sequence and returns the first image.

pub fn first_frame(&self) -> &Frame<P>

Returns a reference to the first frame in the image sequence.