Struct ril::EncoderMetadata

pub struct EncoderMetadata<P: Pixel> {
    pub width: u32,
    pub height: u32,
    pub sequence: Option<(usize, LoopCount)>,
    pub color_type: ColorType,
    pub bit_depth: u8,
    pub palette: Option<Box<[P::Color]>>,
}

Manually configured encoder metadata. This is used to provide fine-grained control over the encoder. Not all encoders will use/consider all of these options.

See Also

[Encoder::prepare] for more information on how to use this.

Encoder::encode_static to encode a static Image.

Encoder::encode_sequence to encode an ImageSequence.

Fields
width: u32

The width of the image.

height: u32

The height of the image.

sequence: Option<(usize, LoopCount)>

Information regarding whether the image is static (not animated) or not.

If Some, the image is animated and holds (frame_count, loop_count).

If None, the image is considered static.

Note

If you do not know the frame count or loop count, substitute 0 for the frame count and LoopCount::Infinite for the loop count. All encoders should be able to “grow” this information as they encode more frames. (This is the default value.)

color_type: ColorType

The color type of the image.

bit_depth: u8

The bit depth of the pixels in the image.

palette: Option<Box<[P::Color]>>

The palette of the image.

Implementations

impl<P: Pixel> EncoderMetadata<P>

pub const fn new(width: u32, height: u32) -> Self

Creates a new encoder metadata with default options given the width, height, and pixel type.

Example

use ril::{EncoderMetadata, Rgba};

let metadata = EncoderMetadata::<Rgba>::new(256, 256);

pub const fn with_width(self, width: u32) -> Self

Sets the width of the image.

pub const fn with_height(self, height: u32) -> Self

Sets the height of the image.

pub const fn with_sequence( self, frame_count: usize, loop_count: LoopCount ) -> Self

Sets sequence information regarding the image.

pub fn with_frame_count(self, frame_count: usize) -> Self

Anticipates the frame count of the image.

pub fn with_loop_count(self, loop_count: LoopCount) -> Self

Anticipates the loop count of the image.

pub const fn with_color_type(self, color_type: ColorType) -> Self

Sets the color type of the image.

pub const fn with_bit_depth(self, bit_depth: u8) -> Self

Sets the bit depth of the pixels in the image.

pub const fn with_config<C: Default>( self, config: C ) -> EncoderMetadataWithConfig<C, P>

Sets the configuration parameters for the specific encoder.

pub fn with_palette(self, palette: impl Deref<Target = [P::Color]>) -> Self

Sets the palette of the image.