Struct ril::Frame

pub struct Frame<P: Pixel> { /* private fields */ }

Represents a frame in an image sequence. It encloses an Image and extra metadata about the frame.

Implementations

impl<P: Pixel> Frame<P>

pub fn from_image(image: Image<P>) -> Self

Creates a new frame with the given image and default metadata.

pub fn set_delay(&mut self, delay: Duration)

Sets the frame delay to the given duration in place.

pub const fn with_delay(self, delay: Duration) -> Self

Takes this frame and sets the frame delay to the given duration.

pub fn set_disposal(&mut self, disposal: DisposalMethod)

Sets the disposal method for this frame in place.

pub const fn with_disposal(self, disposal: DisposalMethod) -> Self

Takes this frame and sets the disposal method for this frame when transitioning to the next.

pub const fn image(&self) -> &Image<P>

Returns a reference to the image this frame contains.

pub fn map_image<T: Pixel>(
self,
f: impl FnOnce(Image<P>) -> Image<T>
) -> Frame<T>

Maps the inner image to the given function.

pub fn image_mut(&mut self) -> &mut Image<P>

Returns a mutable reference to the image this frame contains.

pub fn into_image(self) -> Image<P>

Consumes this frame returning the inner image it represents.

pub const fn delay(&self) -> Duration

Returns the delay duration for this frame.

pub const fn disposal(&self) -> DisposalMethod

Returns the disposal method for this frame.