Struct ril::L

pub struct L(pub u8);

Represents an L, or luminance pixel that is stored as only one single number representing how bright, or intense, the pixel is.

This can be thought of as the “unit channel” as this represents only a single channel in which other pixel types can be composed of.

Tuple Fields
0: u8

The luminance value of the pixel, between 0 and 255.

Implementations

impl L

pub const fn new(l: u8) -> Self

Creates a new L pixel with the given luminance value.

pub const fn value(&self) -> u8

Returns the luminance value of the pixel.