Struct ril::PalettedRgb

pub struct PalettedRgb<'p> {
    pub index: u8,
    /* private fields */
}

Represents a paletted pixel, holding an index to a palette of Rgb colors represented as a &'p [Rgb], where 'p is the lifetime of the palette.

Fields
index: u8