Struct ril::TextLayout

pub struct TextLayout<'a, P: Pixel> { /* private fields */ }

Represents a high-level text layout that can layout text segments, maybe with different fonts.

This is a high-level layout that can be used to layout text segments. It can be used to layout text segments with different fonts and styles, and has many features over TextSegment such as text anchoring, which can be useful for text alignment. This also keeps track of font metrics, meaning that unlike TextSegment, this can be used to determine the width and height of text before rendering it.

This is less efficient than TextSegment and you should use TextSegment if you don’t need any of the features TextLayout provides.

Note

This is does not implement Clone and therefore it is not cloneable! Consider using TextSegment if you require cloning functionality.

Implementations

impl<'a, P: Pixel> TextLayout<'a, P>

pub fn new() -> Self

Creates a new text layout with default settings.

pub fn with_position(self, x: u32, y: u32) -> Self

Sets the position of the text layout.

This must be set before adding any text segments!

pub fn with_wrap(self, wrap: WrapStyle) -> Self

Sets the wrapping style of the text. Make sure to also set the wrapping width using [with_width] for wrapping to work.

This must be set before adding any text segments!

pub fn with_width(self, width: u32) -> Self

Sets the wrapping width of the text. This does not impact Self::width and Self::dimensions.

This must be set before adding any text segments!

pub fn push_segment(&mut self, segment: &TextSegment<'a, P>)

Adds a text segment to the text layout.

pub fn with_segment(self, segment: &TextSegment<'a, P>) -> Self

Takes this text layout and returns it with the given text segment added to the text layout. Useful for method chaining.

pub fn push_basic_text( &mut self, font: &'a Font, text: impl AsRef<str>, fill: P, )

Adds basic text to the text layout. This is a convenience method that creates a TextSegment with the given font, text, and fill and adds it to the text layout.

The size of the text is determined by the font’s optimal size.

Note

The overlay mode is set to OverlayMode::Merge and not the image’s overlay mode, since anti-aliasing is funky with the replace overlay mode.

pub fn with_basic_text( self, font: &'a Font, text: impl AsRef<str>, fill: P, ) -> Self

Takes this text layout and returns it with the given basic text added to the text layout. Useful for method chaining.

Note

The overlay mode is set to OverlayMode::Merge and not the image’s overlay mode, since anti-aliasing is funky with the replace overlay mode.

See Also

push_basic_text

pub const fn with_horizontal_anchor(self, anchor: HorizontalAnchor) -> Self

Sets the horizontal anchor of the text. The horizontal anchor determines where the x position of the text is anchored.

pub const fn with_vertical_anchor(self, anchor: VerticalAnchor) -> Self

Sets the vertical anchor of the text. The vertical anchor determines where the y position of the text is anchored.

pub const fn with_align(self, align: TextAlign) -> Self

Sets the horizontal text alignment. This determines how text is aligned within its bounding box (which is determined by Self::width and Self::height).

pub const fn centered(self) -> Self

Sets the horizontal anchor and vertial anchor of the text to be centered. This makes the position of the text be the center as opposed to the top-left corner.

pub fn width(&self) -> u32

Returns the width of the text. This is a slightly expensive operation and is not a simple getter.

If you want both width and height, use dimensions.

pub fn height(&self) -> u32

Returns the height of the text.

If you want both width and height, use dimensions.

pub fn dimensions(&self) -> (u32, u32)

Returns the width and height of the text. This is a slightly expensive operation and should be used sparingly - it is not a simple getter.

pub fn bounding_box(&self) -> (u32, u32, u32, u32)

Returns the bounding box of the text. Left and top bounds are inclusive; right and bottom bounds are exclusive.