Struct ril::TextSegment

pub struct TextSegment<'a, P: Pixel> {
    pub position: (u32, u32),
    pub width: Option<u32>,
    pub text: String,
    pub font: &'a Font,
    pub fill: P,
    pub overlay: OverlayMode,
    pub size: f32,
    pub wrap: WrapStyle,
}

Represents a text segment that can be drawn.

See TextLayout for a more robust implementation that supports rendering text with multiple styles. This type is for more simple and lightweight usages.

Additionally, accessing metrics such as the width and height of the text cannot be done here, but can be done in TextLayout since it keeps a running copy of the layout. Use TextLayout if you will be needing to calculate the width and height of the text. Additionally, TextLayout supports text anchoring, which can be used to align text.

If you need none of these features, text segments should be used in favor of being much more lightweight.

Note that TextLayout is not cloneable while text segments are, which is one advantage of using this over TextLayout.

Fields
position: (u32, u32)

The position the text will be rendered at. Ignored if this is used in a TextLayout.

width: Option<u32>

The width of the text box. If this is used in a TextLayout, this is ignored and TextLayout::with_width is used instead. This is used for text wrapping and wrapping only.

text: String

The content of the text segment.

font: &'a Font

The font to use to render the text.

fill: P

The fill color the text will be in.

overlay: OverlayMode

The overlay mode of the text. Note that anti-aliasing is still a bit funky with OverlayMode::Replace, so it is best to use OverlayMode::Merge for this, which is the default.

size: f32

The size of the text in pixels.

wrap: WrapStyle

The wrapping style of the text. Note that text will only wrap if [width] is set. If this is used in a TextLayout, this is ignored and TextLayout::with_wrap is used instead.

Implementations

impl<'a, P: Pixel> TextSegment<'a, P>

pub fn new(font: &'a Font, text: impl AsRef<str>, fill: P) -> Self

Creates a new text segment with the given text, font, and fill color. The text can be anything that implements AsRef<str>.

If this is used to be directly drawn (as opposed to in a TextLayout), the position is set to (0, 0) by default. Use with_position to set the position.

The size defaults to the fontโs optimal size. You can override this by using the with_size method.

pub const fn with_position(self, x: u32, y: u32) -> Self

Sets the position of the text segment. Ignored if this segment is used in a TextLayout.

pub const fn with_size(self, size: f32) -> Self

Sets the size of the text segment.

pub const fn with_overlay_mode(self, mode: OverlayMode) -> Self

Sets the overlay mode of the text segment.

pub const fn with_width(self, width: u32) -> Self

Sets the width of the text segment, used for text wrapping. If this is used in a TextLayout, this is ignored and TextLayout::width is used instead.

pub const fn with_wrap(self, wrap: WrapStyle) -> Self

Sets the wrapping style of the text segment. If this is used in a TextLayout, this is ignored and [TextLayout::wrap] is used instead.