ril

Struct Vector

pub struct Vector<const N: usize, T>(/* private fields */);

A mathematical vector represented using a fixed-sized array with elements of type T.

Implementations

impl<const N: usize, T> Vector<N, T>

pub const fn new(elements: [T; N]) -> Self

Creates a new vector with the given elements.

pub const fn of(value: T) -> Self
where T: Copy,

Creates a new vector filled with the given value.

pub const fn elements(&self) -> &[T; N]

Returns the elements of the vector.

pub fn get(&self, index: usize) -> T
where T: Copy,

Returns the element at the given index.

Panics

Panics if the index is out of bounds.

pub fn merge<U: Copy, R: Copy + Zero>( &self, other: Vector<N, U>, f: impl Fn(T, U) -> R, ) -> Vector<N, R>
where T: Copy,

Performs the element-wise operation f on this vector and the given vector other.

pub fn map<R: Copy + Zero>(self, f: impl Fn(T) -> R) -> Vector<N, R>
where T: Copy,

Performs the operation f on each element of this vector.

pub fn dot<U: Copy, R>(&self, other: Vector<N, U>) -> R
where T: Copy + Mul<U, Output = R>, R: Copy + Zero + Add<Output = R>,

Computes the dot product of this vector and the given vector other.

pub fn norm(&self) -> T
where T: Copy + Zero + Add<Output = T> + Mul<Output = T>,

Computes the norm (magnitude squared) of this vector.

pub fn magnitude(&self) -> T
where T: Copy + Zero + Add<Output = T> + Mul<Output = T> + Float,

Computes the magnitude (length) of this vector.

pub fn sum(&self) -> T
where T: Copy + Zero + Add<Output = T>,

Computes the sum of all elements in this vector.

impl<T> Vector<3, T>

pub fn cross<U: Copy, R>(&self, other: Vector<3, U>) -> Vector<3, R>
where T: Copy + Mul<U, Output = R>, R: Copy + Sub<R, Output = R>,

Computes the cross product of this vector and the given vector other.

impl<const N: usize, T: Copy + ConstZero> Vector<N, T>

pub const fn zero() -> Self

Creates a new vector with all elements initialized to zero.

impl<const N: usize, T: Copy + ConstOne> Vector<N, T>

pub const fn one() -> Self

Creates a new vector with all elements initialized to one.