ril

Module text

Implements the font/text rasterizing and layout interface.

Structs

Font

Represents a single font along with its alternatives used to render text. Currently, this supports TrueType and OpenType fonts.

TextLayout

Represents a high-level text layout that can layout text segments, maybe with different fonts.

TextSegment

Represents a text segment that can be drawn.

Enums

HorizontalAnchor

Represents where text is anchored horizontally.

TextAlign

Represents how to align text horizontally within its bounding box.

VerticalAnchor

Represents where text is anchored vertically.

WrapStyle

Determines how text should be wrapped.