ril

Trait IntoVector

pub trait IntoVector<const N: usize> {
    type Element: Sized;

    // Required method
    fn into_vector(self) -> Vector<N, Self::Element>;
}

Vectorizable type trait. This trait is used to indicate that a type (usually a pixel) can be represented as a statically-sized vector.

Required Associated Types

type Element: Sized

The type of each element in the vector representation of this type.

Required Methods

fn into_vector(self) -> Vector<N, Self::Element>

Converts this type into a vector of its elements.

Returns

A vector containing the elements of this type.