Trait ril::Pixel

pub trait Pixel: Copy + Clone + Debug + Default + PartialEq + Eq + Hash {
    type Subpixel: Copy + Into<usize>;
    type Color: Pixel;
    type Data: IntoIterator<Item = u8> + AsRef<[u8]>;

    const COLOR_TYPE: ColorType;
    const BIT_DEPTH: u8;

    fn inverted(&self) -> Self;
    fn map_subpixels<F, A>(self, f: F, a: A) -> Self
    where
        F: Fn(Self::Subpixel) -> Self::Subpixel,
        A: Fn(Self::Subpixel) -> Self::Subpixel;
    fn from_bytes(bytes: &[u8]) -> Self;
    fn as_bytes(&self) -> Self::Data;
    fn merge_with_alpha(self, other: Self, alpha: u8) -> Self;
    fn force_into_rgb(self) -> Rgb;
    fn force_into_rgba(self) -> Rgba;

    fn color_type(&self) -> ColorType { ... }
    fn luminance(&self) -> u8
    where
        Self: Into<L>,
    { ... }
    fn from_raw_parts(
        color_type: ColorType,
        bit_depth: u8,
        data: &[u8]
    ) -> Result<Self> { ... }
    fn from_raw_parts_paletted<P: Pixel>(
        color_type: ColorType,
        bit_depth: u8,
        data: &[u8],
        palette: Option<&[P]>
    ) -> Result<Self> { ... }
    fn from_arbitrary_palette<P: Pixel>(
        palette: &[P],
        index: usize
    ) -> Result<Self> { ... }
    fn merge(self, other: Self) -> Self { ... }
    fn overlay(self, other: Self, mode: OverlayMode) -> Self { ... }
    fn overlay_with_alpha(self, other: Self, mode: OverlayMode, alpha: u8) -> Self { ... }
    fn from_dynamic(dynamic: Dynamic) -> Self { ... }

}

Represents any type of pixel in an image.

Generally speaking, the values enclosed inside of each pixel are designed to be immutable.

Required Associated Types

type Subpixel: Copy + Into<usize>

The type of a single component in the pixel.

type Color: Pixel

The resolved color type of the palette. This is Self for non-paletted pixels.

type Data: IntoIterator<Item = u8> + AsRef<[u8]>

The iterator type this pixel uses.

Required Associated Constants

const COLOR_TYPE: ColorType

The color type of the pixel.

const BIT_DEPTH: u8

The bit depth of the pixel.

Required Methods

fn inverted(&self) -> Self

Returns the inverted value of this pixel.

This does not independently invert the alpha value, instead you may need to split the image and isolate the alpha channel into an Image, invert it, then merge the bands.

Why not invert alpha?

In most cases the alpha channel is favored to be not inverted.

In the case that a pixel is completely transparent, this behavior is almost always the case. The user cannot actually see the color of the pixel, so inverting it to a seemingly unknown color could be unexpected.

For translucent pixels, in most cases the expected behavior is to only invert the color shown for the translucent pixel and not change the alpha itself.

Finally, the most obvious reasoning is that for fully opaque pixels, those pixels will become fully transparent which is obviously not favored.

fn map_subpixels<F, A>(self, f: F, a: A) -> Selfwhere
F: Fn(Self::Subpixel) -> Self::Subpixel,
A: Fn(Self::Subpixel) -> Self::Subpixel,

Maps the pixel’s components and returns a new pixel with the mapped components.

Alpha is intentionally mapped separately. If no alpha component exists, the alpha function is ignored.

fn from_bytes(bytes: &[u8]) -> Self

Creates this pixel from a raw bytes. This is used internally and is unchecked - it panics if the data is not of the correct length.

fn as_bytes(&self) -> Self::Data

Turns this pixel into bytes.

fn merge_with_alpha(self, other: Self, alpha: u8) -> Self

Merges this pixel with the given overlay pixel, where the alpha of the overlay pixel is known. This is used in anti-aliasing.

fn force_into_rgb(self) -> Rgb

Converts this pixel into RGB despite its type. This can panic on some pixel types, you must be sure this pixel is able to be converted into RGB before using this.

You should use Rgb::from or TrueColor::into_rgb as they are safer methods, checked at compile time. This is primarily used for internal purposes, for example when an encoder can guarantee that a pixel is convertable into RGB.

Panics

If the pixel is not convertable into RGB.

fn force_into_rgba(self) -> Rgba

Converts this pixel into RGBA despite its type. This can panic on some pixel types, you must be sure this pixel is able to be converted into RGBA before using this.

You should use Rgba::from or TrueColor::into_rgb as they are safer methods, checked at compile time. This is primarily used for internal purposes, for example when an encoder can guarantee that a pixel is convertable into RGBA.

Panics

If the pixel is not convertable into RGBA.

Provided Methods

fn color_type(&self) -> ColorType

Resolves the color type of this pixel at runtime. This is used for dynamic color types. If you are certain the pixel is not dynamic, you can use the Self::COLOR_TYPE constant instead.

fn luminance(&self) -> u8where
Self: Into<L>,

The luminance of the pixel.

fn from_raw_parts(
color_type: ColorType,
bit_depth: u8,
data: &[u8]
) -> Result<Self>

Creates this pixel from the given color type, bit depth, and data. This may require a lossy conversion.

Errors

If the color type is not supported by the pixel type.

An error occurs when trying to convert the data to the pixel type.

fn from_raw_parts_paletted<P: Pixel>(
color_type: ColorType,
bit_depth: u8,
data: &[u8],
palette: Option<&[P]>
) -> Result<Self>

Creates this pixel from the given color type, bit depth, data, and possibly a color palette. This may require a lossy xonversion.

A palette should be supplied if the color type is paletted, else None.

Errors

If the color type is not supported by the pixel type.

An error occurs when trying to convert the data to the pixel type.

fn from_arbitrary_palette<P: Pixel>(palette: &[P], index: usize) -> Result<Self>

Creates this pixel from the given palette and index, but the conversion is done at runtime.

Errors

The pixel index is invalid/out of bounds.

If the color type is not supported by the pixel type.

An error occurs when trying to convert the data to the pixel type.

fn merge(self, other: Self) -> Self

Merges this pixel with the given overlay pixel, taking into account alpha.

fn overlay(self, other: Self, mode: OverlayMode) -> Self

Overlays this pixel with the given overlay pixel, abiding by the given overlay mode.

fn overlay_with_alpha(self, other: Self, mode: OverlayMode, alpha: u8) -> Self

Overlays this pixel with the given overlay pixel, abiding by the given overlay mode with the given alpha.

This is used in anti-aliasing.

fn from_dynamic(dynamic: Dynamic) -> Self

Creates this pixel from any dynamic pixel, dynamically at runtime. Different from the From/Into traits.