ril::vector

Trait FromVector

pub trait FromVector<const N: usize> {
    type Element: Sized;

    // Required method
    fn from_vector(vector: Vector<N, Self::Element>) -> Self;
}

Vectorizable type trait. This trait is used to indicate that a type can be constructed from a vector.

Required Associated Types

type Element: Sized

The type of each element in the vector representation of this type.

Required Methods

fn from_vector(vector: Vector<N, Self::Element>) -> Self

Constructs a new instance from the given vector.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.