Crate ril

The Rust Imaging Library. A high-level image processing crate for Rust.

Re-exports

pub use draw::Border;

pub use draw::BorderPosition;

pub use draw::Draw;

pub use draw::Ellipse;

pub use draw::Paste;

pub use draw::Rectangle;

pub use encode::Decoder;

pub use encode::DynamicFrameIterator;

pub use encode::Encoder;

pub use encode::FrameIterator;

pub use error::Error;

pub use error::Result;

pub use image::Banded;

pub use image::Image;

pub use image::ImageFormat;

pub use image::OverlayMode;

pub use pixel::Alpha;

pub use pixel::BitPixel;

pub use pixel::Dynamic;

pub use pixel::Pixel;

pub use pixel::Rgb;

pub use pixel::Rgba;

pub use pixel::L;

pub use sequence::DisposalMethod;

pub use sequence::Frame;

pub use sequence::ImageSequence;

pub use sequence::LoopCount;

Modules

draw

encode

encodings

error

image

pixel

prelude

The crate prelude exports. Importing this with a wildcard will import most items from RIL that can be useful for image processing, along with bringing crucial traits into scope.

sequence

Enums

ResizeAlgorithm

A filtering algorithm that is used to resize an image.