Module ril::draw

Encloses most drawing implementations and drawable objects.

Structs

Border

Represents a shape border.

Ellipse

An ellipse, which could be a circle.

Line

A line.

Paste

Pastes or overlays an image on top of another image.

Polygon

A polygon.

Rectangle

A rectangle.

SolidFill

Represents a solid color fill.

Enums

BorderPosition

Represents whether a border is inset, outset, or if it lays in the center.

Traits

Draw

A common trait for all objects able to be drawn on an image.

Fill

Handles the actual filling of a shape. See IntoFill for more information.

IntoFill

Any fill type that can be used to fill a shape, i.e. solid colors or gradients.

Type Definitions

BoundingBox