Trait ril::draw::Draw

pub trait Draw<P: Pixel> {
    fn draw(&self, image: &mut Image<P>);
}

A common trait for all objects able to be drawn on an image.

Whether or not to implement this trait is more or less a matter of semantics.

Required Methods

fn draw(&self, image: &mut Image<P>)

Draws the object to the given image.