Trait ril::encode::FrameIterator

pub trait FrameIterator<P: Pixel>: Iterator<Item = Result<Frame<P>>> {
    fn len(&self) -> u32;
    fn loop_count(&self) -> LoopCount;
    fn into_sequence(self) -> Result<ImageSequence<P>>;
}

Represents the lazy decoding of frames from an encoded image sequence, such as an animated image.

See Also

ImageSequence

Frame

Required Methods

fn len(&self) -> u32

Returns the number of frames in the sequence.

This does not consume any frames as this data is usually known from the very beginning of decoding.

fn loop_count(&self) -> LoopCount

Returns the amount of times this sequence will loop over itself.

fn into_sequence(self) -> Result<ImageSequence<P>>

Collects all frames in this iterator and turns it into a high level ImageSequence.

If any frame fails, that error is returned.