Enum ril::error::Error

pub enum Error {
    InvalidHexCode(String),
    InvalidPaletteIndex,
    InvalidExtension(OsString),
    EncodingError(String),
    DecodingError(String),
    FontError(&'static str),
    UnknownEncodingFormat,
    IncompatibleImageData {
        width: u32,
        height: u32,
        received: usize,
    },
    UnsupportedColorType,
    IOError(Error),
    EmptyImageError,
}

Represents an error that occurs within the crate.

Variants

InvalidHexCode(String)

An invalid hex code was provided when trying to parse a hex value.

InvalidPaletteIndex

Received an invalid palette index.

InvalidExtension(OsString)

An invalid extension was provided when trying to resolve an image’s encoding format from a file extension.

Note

This is not an error that occurs when the file extension is not recognized, or is an unknown image extension. This occurs if the OsStr fails conversion to a native &str. In the case of this, [ImageFormat::Unknown] is used instead.

EncodingError(String)

Failed to encode an image.

DecodingError(String)

Invalid data was encountered when an image, usually because it is corrupted.

Errors can differ across encodings, so the inner &'static str here is nothing more than an error message.

FontError(&'static str)

An error occured while trying to render or rasterize a font.

UnknownEncodingFormat

No encoding format could be inferred for the given image.

IncompatibleImageData

Fields

width: u32

height: u32

received: usize

An image received data incompatible with the image’s dimensions.

UnsupportedColorType

Received an unsupported color type when trying to create a pixel from raw data.

This occurs when the color type is not supported by the pixel type. This is almost always fixed by switching the pixel type to [Dynamic] then using [Image::convert] to convert the image into your desired type.

IOError(Error)

An error occured when trying to read a file or when trying to write to a file.

EmptyImageError

Tried to encode an empty image, or an image without data.