Trait ril::fill::IntoFill

pub trait IntoFill: Clone {
    type Pixel: Pixel;
    type Fill: Fill<Self::Pixel>;

    // Required method
    fn into_fill(self) -> Self::Fill;
}

Any fill type that can be used to fill a shape, i.e. solid colors or gradients.

For solid colors, this trait is implemented for all Pixel types as SolidFill.

Required Associated Types

type Pixel: Pixel

The pixel type of the fill.

type Fill: Fill<Self::Pixel>

The fill type.

Required Methods

fn into_fill(self) -> Self::Fill

Converts the fill into a fill type.

Object Safety

This trait is not object safe.