ril::filter

Trait Filter

pub trait Filter {
    type Input: Pixel;
    type Output: Pixel;

    // Required method
    fn apply_pixel(
        &self,
        image: &Image<Self::Input>,
        x: u32,
        y: u32,
        pixel: Self::Input,
    ) -> Self::Output;
}

An image filter than can be lazily applied to an image or a filtered image.

Filters make the following guarantees:

They preserve the image dimensions.

They modify single pixels at a time. They can use surrounding pixels for context, but they cannot modify them.

Required Associated Types

type Input: Pixel

The pixel type of the input image.

type Output: Pixel

The pixel type of the output image.

Required Methods

fn apply_pixel( &self, image: &Image<Self::Input>, x: u32, y: u32, pixel: Self::Input, ) -> Self::Output

Applies the filter to the given pixel.