Struct ril::gradient::RadialGradient

pub struct RadialGradient<P: Pixel> {
    pub position: GradientPosition,
    pub cover: RadialGradientCover,
    pub colors: Vec<(P, f64)>,
    pub interpolation: Interpolation,
    pub blend_mode: BlendMode,
}

A radial gradient.

Fields
position: GradientPosition

The position of the center of the radial gradient (where the radial gradient “radiates” from).

cover: RadialGradientCover

How the gradient should cover the bounding box.

colors: Vec<(P, f64)>

A Vec of colors and their positions in the gradient, represented as (color, position) where position is a value in the range [0.0, 1.0].

Normalization of positions

During building of this struct, there might be some positions that are nan which represent positions that will be normalized later. For example, [0.0, nan, 1.0] is normalized to [0.0, 0.5, 1.0] because 0.5 is the midpoint between 0.0 and 1.0.

Similarly, [0.0, nan, nan, nan, 1.0] is normalized to [0.0, 0.25, 0.5, 0.75, 1.0] because they evenly distribute between 0.0 and 1.0.

Normalization of endpoints

If the first position is nan, it will be normalized to 0.0. If the last position is nan, it will be normalized to 1.0.

interpolation: Interpolation

The interpolation mode to use when rendering the gradient. Defaults to Interpolation::Linear.

blend_mode: BlendMode

The blending mode to use when rendering the gradient. Defaults to BlendMode::LinearRgb. If the gradient looks off or some colors are weirdly balanced, trying different blend modes here could help.

Implementations

impl<P: Pixel> RadialGradient<P>

pub fn new() -> Self

Creates a new empty radial gradient.

pub const fn with_position(self, position: GradientPosition) -> Self

Sets the position of the center of the radial gradient (where the radial gradient “radiates” from).

pub const fn with_cover(self, cover: RadialGradientCover) -> Self

Sets how the gradient should cover the bounding box.

pub const fn with_interpolation(self, interpolation: Interpolation) -> Self

Sets the interpolation mode to use when rendering the gradient.

pub const fn with_blend_mode(self, blend_mode: BlendMode) -> Self

Sets the blending mode to use when rendering the gradient.

pub fn with_start_color(self, color: P) -> Self

Sets the start color of the gradient. This will be rendered at the position 0.0.

Note

This uses insert instead of push to ensure that the start color is always at the beginning of the gradient. This means other colors will be shifted to the right.

pub fn with_end_color(self, color: P) -> Self

Sets the end color of the gradient. This will be rendered at the position 1.0.

pub fn push_color_at(&mut self, position: f64, color: P)

Adds a color to the gradient at the specified position in place.

Panics

If the position is outside of the range [0.0, 1.0]. For auto-normalized positions, see Self::push_color.

pub fn with_color_at(self, position: f64, color: P) -> Self

Takes this gradient and adds a color to the gradient at the specified position.

Panics

If the position is outside of the range [0.0, 1.0]. For auto-normalized positions, see Self::with_color.

pub fn push_color(&mut self, color: P)

Adds a color to the gradient, automatically calculating its position. See the documentation for [Self.colors] for more information of how colors are normalized.

See Also

Self::push_color_at for adding a color at a specific position.

pub fn with_color(self, color: P) -> Self

Takes this gradient and adds a color to the gradient, automatically calculating its position. See the documentation for [Self.colors] for more information of how colors are normalized.

See Also

Self::with_color_at for adding a color at a specific position.

pub fn extend_with_positions<I: IntoIterator<Item = (P, f64)>>( &mut self, iter: I )

Extends the colors and positions of this gradient with those specified in the given iterator of tuples represented as (color, position).

Panics

If any of the positions are outside of the range [0.0, 1.0]. For auto-normalized positions, see Self::extend.

pub fn extend<I: IntoIterator<Item = P>>(&mut self, iter: I)

Extends the colors of this gradient with those specified in the given iterator. The positions of the colors will be automatically calculated. See the documentation for [Self.colors] for more information of how colors are normalized.

See Also

Self::extend_with_positions for adding colors at specific positions.