Module ril::pixel

Encloses pixel-related traits and pixel type implementations.

Structs

BitPixel

Represents a single-bit pixel that represents either a pixel that is on or off.

L

Represents an L, or luminance pixel that is stored as only one single number representing how bright, or intense, the pixel is.

NoOp

A pixel type that does and stores nothing. This pixel type is useless and will behave weirdly with your code. This is usually only used for internal or polyfill purposes.

NoOpSubpixel

Extension of NoOp, used for internal purposes only. This is a ZST that implements Into<u8>.

PalettedRgb

Represents a paletted pixel, holding an index to a palette of Rgb colors represented as a &'p [Rgb], where 'p is the lifetime of the palette.

PalettedRgba

Represents a paletted pixel, holding an index to a palette of Rgba colors represented as a &'p [Rgba], where 'p is the lifetime of the palette.

Rgb

Represents an RGB pixel.

Rgba

Represents an RGBA pixel.

Enums

Dynamic

Represents a pixel type that is dynamically resolved.

DynamicSubpixel

Represents a subpixel of a dynamic pixel.

Traits

Alpha

Represents a pixel that supports alpha, or transparency values.

Paletted

A trait representing a paletted pixel. Pixel::Subpixel is the type of the palette index.

Pixel

Represents any type of pixel in an image.

TrueColor

A trait representing all pixels that can be safely represented as either RGB or RGBA true color.