Module ril::pixel

Encloses pixel-related traits and pixel type implementations.

Structs

BitPixel

Represents a single-bit pixel that represents either a pixel that is on or off.

L

Represents an L, or luminance pixel that is stored as only one single number representing how bright, or intense, the pixel is.

Rgb

Represents an RGB pixel.

Rgba

Represents an RGBA pixel.

Enums

Dynamic

Represents a pixel type that is dynamically resolved.

Traits

Alpha

Represents a pixel that supports alpha, or transparency values.

Pixel

Represents any type of pixel in an image.