Trait ril::pixel::Pixel

pub trait Pixel: Copy + Clone + Default + PartialEq + Eq {
    fn alpha(&self) -> u8;
    fn inverted(&self) -> Self;
    fn from_pixel_data(data: PixelData) -> Result<Self>;

    fn luminance(&self) -> u8
    where
        Self: Into<L>,
    { ... }
}

Represents any type of pixel in an image.

Required Methods

fn alpha(&self) -> u8

Returns the alpha, or opacity level of the pixel.

This is a value between 0 and 255. 0 is completely transparent, and 255 is completely opaque.

fn inverted(&self) -> Self

Returns the inverted value of this pixel.

This does not independently invert the alpha value, instead you may need to split the image and isolate the alpha channel into an Image, invert it, then merge the bands.

Why not invert alpha?

In most cases the alpha channel is favored to be not inverted.

In the case that a pixel is completely transparent, this behavior is almost always the case. The user cannot actually see the color of the pixel, so inverting it to a seemingly unknown color could be unexpected.

For translucent pixels, in most cases the expected behavior is to only invert the color shown for the translucent pixel and not change the alpha itself.

Finally, the most obvious reasoning is that for fully opaque pixels, those pixels will become fully transparent which is obviously not favored.

fn from_pixel_data(data: PixelData) -> Result<Self>

Creates this pixel from raw data.

Provided Methods

fn luminance(&self) -> u8 where
Self: Into<L>,

The luminance of the pixel.