Module ril::prelude

The crate prelude exports. Importing this with a wildcard will import most items from RIL that can be useful for image processing, along with bringing crucial traits into scope.

Example

use ril::prelude::*;

// Prelude imported Image and Rgb
let image = Image::new(100, 100, Rgb::new(255, 0, 0));
// Prelude imported the Banded trait
let (r, g, b) = image.bands();

Re-exports

pub use super::Alpha;

pub use super::Banded;

pub use super::BitPixel;

pub use super::Border;

pub use super::BorderPosition;

pub use super::ColorType;

pub use super::DisposalMethod;

pub use super::Draw;

pub use super::Dynamic;

pub use super::DynamicFrameIterator;

pub use super::DynamicSubpixel;

pub use super::Ellipse;

pub use super::Fill;

pub use super::Frame;

pub use super::FrameIterator;

pub use super::Image;

pub use super::ImageFormat;

pub use super::ImageSequence;

pub use super::IntoFill;

pub use super::Line;

pub use super::LoopCount;

pub use super::OverlayMode;

pub use super::Paletted;

pub use super::PalettedRgb;

pub use super::PalettedRgba;

pub use super::Paste;

pub use super::Pixel;

pub use super::Polygon;

pub use super::Rectangle;

pub use super::Rgb;

pub use super::Rgba;

pub use super::TrueColor;

pub use super::L;

pub use super::ResizeAlgorithm;

pub use super::ConicGradient;

pub use super::GradientPosition;

pub use super::LinearGradient;

pub use super::RadialGradient;

pub use super::RadialGradientCover;

pub use super::Font;

pub use super::HorizontalAnchor;

pub use super::TextLayout;

pub use super::TextSegment;

pub use super::VerticalAnchor;

pub use super::WrapStyle;

Enums

GradientBlendMode

Color blending mode

GradientInterpolation

Interpolation mode