Struct ril::quantize::Quantizer

pub struct Quantizer {
    pub palette_size: usize,
    pub gif_optimization: bool,
    pub quality: u8,
    pub fallback_to_lossless: bool,
}

Configuration options regarding behavior of quantization.

Fields
palette_size: usize

The maximum number of colors in the palette. Defaults to 256.

gif_optimization: bool

Whether to optimize the palette for GIF images.

quality: u8

The quality of the output colors. If you plan on using a lossless quantizer like quantize_simple, this value will be disregarded. Otherwise, lower values will produce lower quality colors in a faster time, and higher values will produce higher quality colors in a slower amount of time.

The quality can be any value between 1 and 30. Defaults to 20.

fallback_to_lossless: bool

When attempting lossy quantization and this value is true, the quantizer will check if the amount of unique colors in the image is less than the desired palette size. If it is, the quantizer will fallback to a lossless quantization algorithm.

This does incur a O(n) time complexity when calculating the amount of unique colors in the image, where n is the amount of pixels in the image. This isn’t very significant, but it’s worth noting.

Implementations

impl Quantizer

pub fn new() -> Self

Creates a new [QuantizerConfig] with default settings.

pub const fn with_palette_size(self, palette_size: usize) -> Self

Sets the maximum number of colors in the palette.

pub const fn with_gif_optimization(self, gif_optimization: bool) -> Self

Sets whether to optimize the palette for GIF images.

pub const fn with_quality(self, quality: u8) -> Self

Sets the quality of the output colors. If you plan on using a lossless quantizer like quantize_simple, this value will be disregarded. Otherwise, lower values will produce lower quality colors in a faster time, and higher values will produce higher quality colors in a slower amount of time.

The quality can be any value between 1 and 30.

pub const fn with_fallback_to_lossless(self, fallback_to_lossless: bool) -> Self

Sets whether to fallback to a lossless quantization algorithm when attempting lossy quantization and the amount of unique colors in the image is less than the desired palette size.

This does incur a O(n) time complexity when calculating the amount of unique colors in the image, where n is the amount of pixels in the image. This isn’t very significant, but it’s worth noting.

pub fn quantize<P: TrueColor>(
&self,
pixels: impl AsRef<[P]>
) -> Result<(Vec<P>, Vec<u8>)>

Quantizes the given pixels to a palette of the given size. Returns (palette, image_data).

Behavior

If the quantize feature is disabled (enabled by default), this function will simply run the quantize_simple function (lossless) and return Err if there are more unique colors than the palette size.

Otherwise, this function will always favor a lossy quantization algorithm unless the fallback_to_lossless option is set to true (which is the default) and the amount of unique colors in the image is less than the desired palette size.

Errors

There are more unique colors than the palette size, and the quantize feature is disabled