Struct ril::PalettedRgba

pub struct PalettedRgba<'p> {
    pub index: u8,
    /* private fields */
}

Represents a paletted pixel, holding an index to a palette of Rgba colors represented as a &'p [Rgba], where 'p is the lifetime of the palette.

Fields
index: u8