Trait ril::Fill

pub trait Fill<P: Pixel>: Clone {
    // Required method
    fn get_pixel(&self, x: u32, y: u32) -> P;

    // Provided methods
    fn set_bounding_box(&mut self, _bounding_box: BoundingBox<u32>) { ... }
    fn with_bounding_box(self, bounding_box: BoundingBox<u32>) -> Self
       where Self: Sized { ... }
    fn plot(&self, image: &mut Image<P>, x: u32, y: u32, mode: OverlayMode) { ... }
    fn plot_with_alpha(
        &self,
        image: &mut Image<P>,
        x: u32,
        y: u32,
        mode: OverlayMode,
        alpha: u8
    ) { ... }
}

Handles the actual filling of a shape. See IntoFill for more information.

Required Methods

fn get_pixel(&self, x: u32, y: u32) -> P

Gets the color of the fill at the given coordinates.

Provided Methods

fn set_bounding_box(&mut self, _bounding_box: BoundingBox<u32>)

Sets the bounding box of the fill in place. This is used internally.

fn with_bounding_box(self, bounding_box: BoundingBox<u32>) -> Selfwhere Self: Sized,

Sets the overlay mode of the fill. This is used internally.

fn plot(&self, image: &mut Image<P>, x: u32, y: u32, mode: OverlayMode)

Plots the fill at the given coordinates on the given image.

fn plot_with_alpha( &self, image: &mut Image<P>, x: u32, y: u32, mode: OverlayMode, alpha: u8 )

Plots the fill at the given coordinates on the given image with a custom alpha value.