Trait ril::FrameIterator

pub trait FrameIterator<P: Pixel>: Iterator<Item = Result<Frame<P>>> {
    fn len(&self) -> u32;
    fn loop_count(&self) -> LoopCount;

    fn is_empty(&self) -> bool { ... }
    fn into_sequence(self) -> Result<ImageSequence<P>>
    where
        Self: Sized,
    { ... }
}

Represents the lazy decoding of frames from an encoded image sequence, such as an animated image.

See Also

ImageSequence

Frame

Required Methods

fn len(&self) -> u32

Returns the number of frames in the sequence.

This does not consume any frames as this data is usually known from the very beginning of decoding.

fn loop_count(&self) -> LoopCount

Returns the amount of times this sequence will loop over itself.

Provided Methods

fn is_empty(&self) -> bool

Returns if there are no frames in the sequence. In this case, the image is probably invalid to be encoded again.

fn into_sequence(self) -> Result<ImageSequence<P>>where
Self: Sized,

Collects all frames in this iterator and turns it into a high level ImageSequence. If any frame fails, that error is returned.

Errors

An error occured during decoding one of the frames.