Trait ril::Pixel

pub trait Pixel: Copy + Clone + Default + PartialEq + Eq {
    type Data: IntoIterator<Item = u8>;

    fn inverted(&self) -> Self;
    fn from_pixel_data(data: PixelData) -> Result<Self>;
    fn as_pixel_data(&self) -> PixelData;
    fn as_bytes(&self) -> Self::Data;
    fn merge_with_alpha(self, other: Self, alpha: u8) -> Self;
    fn from_dynamic(dynamic: Dynamic) -> Self;

    fn luminance(&self) -> u8
    where
        Self: Into<L>,
    { ... }
    fn merge(self, other: Self) -> Self { ... }
    fn overlay(self, other: Self, mode: OverlayMode) -> Self { ... }
    fn overlay_with_alpha(
        self, 
        other: Self, 
        mode: OverlayMode, 
        alpha: u8
    ) -> Self { ... }
}

Represents any type of pixel in an image.

Generally speaking, the values enclosed inside of each pixel are designed to be immutable.

Required Associated Types

type Data: IntoIterator<Item = u8>

Required Methods

fn inverted(&self) -> Self

Returns the inverted value of this pixel.

This does not independently invert the alpha value, instead you may need to split the image and isolate the alpha channel into an Image, invert it, then merge the bands.

Why not invert alpha?

In most cases the alpha channel is favored to be not inverted.

In the case that a pixel is completely transparent, this behavior is almost always the case. The user cannot actually see the color of the pixel, so inverting it to a seemingly unknown color could be unexpected.

For translucent pixels, in most cases the expected behavior is to only invert the color shown for the translucent pixel and not change the alpha itself.

Finally, the most obvious reasoning is that for fully opaque pixels, those pixels will become fully transparent which is obviously not favored.

fn from_pixel_data(data: PixelData) -> Result<Self>

Creates this pixel from raw data.

Errors

todo!()

fn as_pixel_data(&self) -> PixelData

Creates raw pixel data from this pixel type.

fn as_bytes(&self) -> Self::Data

Turns this pixel into bytes.

fn merge_with_alpha(self, other: Self, alpha: u8) -> Self

Merges this pixel with the given overlay pixel, where the alpha of the overlay pixel is known. This is used in anti-aliasing.

fn from_dynamic(dynamic: Dynamic) -> Self

Creates this pixel from any dynamic pixel…dynamically at runtime. Different from the From/Into traits.

Provided Methods

fn luminance(&self) -> u8 where
Self: Into<L>,

The luminance of the pixel.

fn merge(self, other: Self) -> Self

Merges this pixel with the given overlay pixel, taking into account alpha.

fn overlay(self, other: Self, mode: OverlayMode) -> Self

Overlays this pixel with the given overlay pixel, abiding by the given overlay mode.

fn overlay_with_alpha(self, other: Self, mode: OverlayMode, alpha: u8) -> Self

Overlays this pixel with the given overlay pixel, abiding by the given overlay mode with the given alpha.

This is used in anti-aliasing.